feat: implement AddMorePhotosForm component and update image upload functionality
