+++ frontend/src/components/PropertyComponents/CreateNewProperty/CreateNewProperty.css
@@ -30,6 +30,14 @@
   max-width: 16rem;
 }
 
+.upload-image-submit-button-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 1rem;
+  width: 100%;
+}
+
 /* @media screen and (min-width: 768px) {
   .form-container {
     padding: 0 3rem;
