+++ frontend/src/components/PropertyComponents/CreateNewProperty/FormInputComponents/NewPropertyComponent/AddressInputComponents/AddressForm.jsx
@@ -1,8 +1,8 @@
 import { useMemo, useEffect } from "react";
 import "./AddressForm.css";
 import GoogleMaps from "./GoogleMaps";
-import useLocationHook from "../../../../../hooks/useLocationHook";
-import Input from "../InputComponents/Input";
+import useLocationHook from "../../../../../../hooks/useLocationHook";
+import Input from "../../InputComponents/Input";
 const AddressForm = ({
   street,
