+++ frontend/src/components/PropertyComponents/CreateNewProperty/FormInputComponents/NewPropertyComponent/PropertyDescription/PropertyDescription.css
